+++ tests/test_ghost_exchange_volume.cpp
@@ -112,11 +112,13 @@ static void test(const testenv::TEnv &t,
     for (const auto &g : gexds) {
         for (const auto &sendc : g.send) {
             BOOST_TEST(((0 <= sendc.value())
-                        && (sendc.value() < grid->local_cells().size())));
+                        && (static_cast<size_t>(sendc.value())
+                            < grid->local_cells().size())));
         }
         for (const auto &recvc : g.recv) {
             BOOST_TEST(((recvc.value() >= 0)
-                        && (recvc.value() < grid->ghost_cells().size())));
+                        && (static_cast<size_t>(recvc.value())
+                            < grid->ghost_cells().size())));
         }
     }
